Add : 상수데이터 적용
