+++ src/components/FeedsArea/FeedsArea.scss
@@ -1,0 +1,120 @@
+@charset "utf-8";
+@import '../../styles/variable.scss';
+
+article.feed {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin-top: 40px;
+  margin-right: 30px;
+  width: 480px;
+  background-color: $bgbox-color;
+  border: 1px solid #e1e1e1;
+  border-radius: 10px;
+
+  div.feedNav {
+    display: flex;
+    align-items: center;
+    position: relative;
+    margin: 15px 10px;
+
+    .feedNavIconName {
+      @include flex-center;
+
+      #feedProfileIcon {
+        width: 30px;
+        border-radius: 50%;
+        border: 2px solid #e1e1e1;
+      }
+
+      .feedProfileName {
+        font-size: 16px;
+        font-weight: 600;
+        margin-left: 5px;
+      }
+    }
+
+    .feedPlusIcon {
+      margin-left: 280px;
+      width: 30px;
+      margin-right: 15px;
+    }
+  }
+
+  div.feedImageSection {
+    width: 480px;
+    height: 500px;
+    overflow: hidden;
+    border-top: 1px solid #e1e1e1;
+    position: relative;
+
+    #feedImg {
+      width: 480px;
+      position: absolute;
+      height: auto;
+      transform: translate(0%, -20%);
+    }
+  }
+
+  div.feedImageResponse {
+    @include flex-space-between;
+    margin: 10px;
+
+    .feedImageResponseInner {
+      display: flex;
+
+      img {
+        height: 30px;
+        width: 30px;
+        margin-right: 15px;
+      }
+    }
+
+    img {
+      height: 30px;
+      width: 30px;
+    }
+  }
+
+  article.commentArea {
+    margin: 0 15px;
+
+    .heartBeatPeople {
+      margin-bottom: 5px;
+
+      .userName {
+        font-weight: 600;
+      }
+    }
+  }
+
+  div.commentBox {
+    border-top: 1px solid #e1e1e1;
+
+    .inputCommentsBox {
+      @include flex-space-between;
+      margin: 20px 15px;
+      padding-left: 2px;
+
+      #commentInput {
+        width: 400px;
+        border: 0px solid;
+        outline: none;
+        font-size: 1em;
+      }
+
+      .commentSubmit {
+        cursor: pointer;
+        font-weight: 600;
+        color: #e1e1e1;
+        font-size: 1em;
+      }
+      .commentSubmit_pass {
+        cursor: pointer;
+        font-weight: 600;
+        color: $point-color;
+        font-size: 1em;
+      }
+    }
+  }
+}
