+++ src/components/Nav/Nav.js
@@ -4,7 +4,7 @@ import "./Nav.scss"
 const Nav = () => {
   return (
     <div>
-      <div className="Navigater"><h1>Nav</h1></div>
+      {/* <div className="Navigater"><h1>Nav</h1></div> */}
     </div>
   );
 };
