+++ src/components/NewComment/NewComment.scss
@@ -1,0 +1,6 @@
+@charset "utf-8";
+@import '../../styles/variable.scss';
+
+.newComment {
+  margin-bottom: 10px;
+}
