+++ src/pages/shinsoojung/Main/Main.js
@@ -1,10 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
-
+import FeedsArea from '../../../components/FeedsArea/FeedsArea'
 import './Main.scss';
 
 const Main = () => {
   const navigate = useNavigate();
+
   const goToLogin = () => {
     navigate('/login-shinsoojung');
   };
@@ -58,8 +59,7 @@ const Main = () => {
       </nav>
       <main>
         <section className="feedScrollArea">
-            <CommentArea />
-
+            <FeedsArea />
         </section>
         <aside className="stickyAsideArea">
           <ul>
@@ -99,36 +99,11 @@ const Main = () => {
               </div>
               <div className="recommendPeopleFollow">팔로잉</div>
             </li>
-            <li className="recommandPeopleBox">
-              <div className="recommandPeopleList">
-                <div className="recommendProfileIcon">
-                  <img
-                    src="images/shinsoojung/Main/myprofile.png"
-                    alt="myProfile"
-                  />
-                </div>
-                <div className="recommendProfileName">
-                  doublejs1226
-                  <div className="recommendProfileSubName">수정</div>
-                </div>
-              </div>
-              <div className="recommendPeopleFollow">팔로잉</div>
-            </li>
-            <li className="recommandPeopleBox">
-              <div className="recommandPeopleList">
-                <div className="recommendProfileIcon">
-                  <img
-                    src="images/shinsoojung/Main/myprofile.png"
-                    alt="myProfile"
-                  />
-                </div>
-                <div className="recommendProfileName">
-                  doublejs1226
-                  <div className="recommendProfileSubName">수정</div>
-                </div>
-              </div>
-              <div className="recommendPeopleFollow">팔로잉</div>
-            </li>
+          </ul>
+          <ul className='systemNotices'> 
+          {SYSTEM_NOTICS.map((notics)=>{
+            return <li key={notics.id} >{notics.name}</li>
+          })}
           </ul>
         </aside>
       </main>
@@ -136,112 +111,13 @@ const Main = () => {
   );
 };
 
-const CommentArea = () => {
-  const [inputValue, setinputValue] = useState('');
-  const [commentList, setcommentList] = useState([]);
-  // const [disabled, setDisabled] = useState(true);
-
-
-  const addComment =(e)=>{
-    e.preventDefault();
-    setcommentList([...commentList,inputValue]);
-    setinputValue('');
-  }
-
-  // const handleBtn = () => {
-  //   inputValue.length > 0
-  //     ? setDisabled(false)
-  //     : setDisabled(true);
-  // };
-  // console.log(inputValue);
-
-  return (
-    <article className="feed">
-            <div className="feedNav">
-              <div className="feedNavIconName">
-                <img
-                  id="feedProfileIcon"
-                  src="images/shinsoojung/Main/myprofile.png"
-                  alt="myProfile"
-                />
-                <div className="feedProfileName">doublejs1226</div>
-              </div>
-              <img
-                className="feedPlusIcon"
-                src="images/shinsoojung/Main/dot.png"
-                alt="feedNavDot"
-              />
-            </div>
-            <div className="feedImageSection">
-              <img
-                src="images/shinsoojung/Main/feed_img1.jpg"
-                alt="feedimg1"
-                id="feedImg"
-              />
-            </div>
-            <div className="feedImageResponse">
-              <div className="feedImageResponseInner">
-                <img
-                  src="images/shinsoojung/Main/heartbeat.jpg"
-                  alt="heartBeatIcon"
-                />
-                <img
-                  src="images/shinsoojung/Main/comment.jpg"
-                  alt="commentIcon"
-                />
-                <img src="images/shinsoojung/Main/dm.jpg" alt="dmIcon" />
-              </div>
-              <img
-                src="images/shinsoojung/Main/cliping.jpg"
-                alt="clipingIcon"
-              />
-            </div>
-      <article className="commentArea">
-        <div className="heartBeatPeople">
-          <p>
-            <span className="userName">doublejs1226</span>님
-            <span className="userName">여러 명</span>이 좋아합니다.
-          </p>
-        </div>
-      <NewComment commentList={commentList}  />
-      </article>
-      <div className="commentBox">
-        <form className="inputCommentsBox">
-          <input
-            onChange={e => {
-              setinputValue(e.target.value);
-            }}
-            value={inputValue}
-            // onKeyUp={handleBtn}
-            id="commentInput"
-            type="text"
-            placeholder="댓글 달기..."
-          />
-          <button
-          onClick={addComment}
-          disabled={inputValue ? false : true}
-          className={inputValue ? 'commentSubmit_pass':'commentSubmit'}>
-          게시</button>          
-        </form>
-      </div>
-    </article>
-  );
-};
-
-
-const NewComment = ({commentList}) => {
-  return (
-    <div className="newComment">
-    {/* 신규 코멘트 영역 */}
-    {commentList.map((data, i) => {
-      return <div data={data} key={i}>{data}</div>
-    })}
-  </div>
-  )
-}
-
-
-
+const SYSTEM_NOTICS = [
+  {id:1, name : "소개"},
+  {id:2, name : "홍보센터"},
+  {id:3, name : "API"},
+  {id:4, name : "개인정보처리방침"},
+  {id:5, name : "약관"} 
+];
 
 
 export default Main;
