+++ src/pages/shinsoojung/Main/Main.scss
@@ -1,31 +1,14 @@
 @charset "utf-8";
 @import '../../../styles/reset.scss';
 @import '../../../styles/common.scss';
-
-@import url('https://fonts.googleapis.com/css2?family=Berkshire+Swash&family=Gothic+A1:wght@400;600;900&display=swap');
-
-
-$bg-color:rgb(249, 249, 249);
-$bgbox-color:#fff;
-$point-color: #6eacf5;
-
-@mixin flex-center {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@mixin flex-space-between {
-  display: flex;
-  justify-content: space-between;
-}
+@import '../../../styles/variable.scss';
 
 nav {
   display: flex;
   justify-content: center;
   position: fixed;
   top: 0;
-  background-color:$bgbox-color;
+  background-color: $bgbox-color;
   z-index: 1;
 
   .homeNavBox {
@@ -39,7 +22,6 @@ nav {
       font-family: 'Berkshire Swash', cursive, serif;
       font-size: 1.7em;
     }
-
 
     div.searchBar {
       position: relative;
@@ -61,7 +43,6 @@ nav {
         border-radius: 5px;
         border: 1px solid #eee;
       }
-
     }
 
     div.navIcon {
@@ -79,11 +60,8 @@ nav {
         border-radius: 50%;
         border: 2px solid #e1e1e1;
       }
-
     }
-
   }
-
 }
 
 /*main start*/
@@ -94,146 +72,14 @@ main {
   margin-top: 80px;
   background-color: $bg-color;
 
-    /* section start */
+  /* section start */
 
   section.feedScrollArea {
     display: flex;
     flex-direction: column;
-
-    article.feed {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      margin-top: 40px;
-      margin-right: 30px;
-      width: 480px;
-      background-color: $bgbox-color;
-      border: 1px solid #e1e1e1;
-      border-radius: 10px;
-
-      div.feedNav {
-        display: flex;
-        align-items: center;
-        position: relative;
-        margin: 15px 10px;
-
-        .feedNavIconName {
-          @include flex-center;
-
-          #feedProfileIcon {
-            width: 30px;
-            border-radius: 50%;
-            border: 2px solid #e1e1e1;
-          }
-
-          .feedProfileName {
-            font-size: 16px;
-            font-weight: 600;
-            margin-left: 5px;
-          }
-
-        }
-
-        .feedPlusIcon {
-          margin-left: 280px;
-          width: 30px;
-          margin-right: 15px;
-        }
-
-      }
-
-      div.feedImageSection {
-        width: 480px;
-        height: 500px;
-        overflow: hidden;
-        border-top: 1px solid #e1e1e1;
-        position: relative;
-
-        #feedImg {
-          width: 480px;
-          position: absolute;
-          height: auto;
-          transform: translate(0%, -20%);
-        }
-      }
-
-      div.feedImageResponse {
-        @include flex-space-between;
-        margin: 10px;
-
-        .feedImageResponseInner {
-          display: flex;
-
-          img {
-            height: 30px;
-            width: 30px;
-            margin-right: 15px;
-          }
-        }
-
-        img {
-          height: 30px;
-          width: 30px;
-        }
-
-      }
-
-      article.commentArea {
-        margin: 0 15px;
-
-        .heartBeatPeople {
-          margin-bottom: 5px;
-
-          .userName {
-            font-weight: 600;
-          }
-
-        }
-
-        .newComment {
-          margin-bottom: 10px;
-        }
-
-      }
-
-      div.commentBox {
-        border-top: 1px solid #e1e1e1;
-
-        .inputCommentsBox {
-          @include flex-space-between;
-          margin: 20px 15px;
-          padding-left: 2px;
-
-          #commentInput {
-            width: 400px;
-            border: 0px solid;
-            outline: none;
-            font-size: 1em;
-          }
-
-          .commentSubmit {
-            cursor: pointer;
-            font-weight: 600;
-            color: #e1e1e1;
-            font-size: 1em;
-          }
-          .commentSubmit_pass {
-            cursor: pointer;
-            font-weight: 600;
-            color: $point-color;
-            font-size: 1em;
-          }
-
-        }
-
-      }
-
-    }
-
   }
 
-    /*section end*/
-
+  /*section end*/
 
   /* aside start */
 
@@ -244,7 +90,6 @@ main {
       @include flex-space-between;
       align-items: center;
       margin-top: 60px;
-
 
       .asideMyProfile {
         display: flex;
@@ -277,7 +122,6 @@ main {
         font-size: 0.9em;
         color: $point-color;
       }
-
     }
 
     li.recommandListTitle {
@@ -294,13 +138,11 @@ main {
         display: flex;
 
         .recommendProfileIcon {
-
           img {
             width: 40px;
             border-radius: 50%;
             border: 2px solid #e1e1e1;
           }
-
         }
 
         .recommendProfileName {
@@ -313,20 +155,30 @@ main {
           font-weight: 300;
           opacity: 0.8;
         }
-
       }
 
       div.recommendPeopleFollow {
         font-size: 0.9em;
         color: $point-color;
       }
-
     }
 
+    ul.systemNotices {
+      display: flex;
+      flex-wrap: wrap;
+      margin-top: 20px;
+      font-size: 1em;
+      color: $basic-font-color;
+      opacity: 0.5;
+
+      li {
+        padding-right: 6px;
+        line-height: 1.5em;
+      }
+    }
   }
 
   /*aside end*/
-
 }
 
 /*main end*/
