+++ src/styles/variable.scss
@@ -1,4 +1,21 @@
 
-body {
-  background-color: #f0f0f0;
+
+@import url('https://fonts.googleapis.com/css2?family=Berkshire+Swash&family=Gothic+A1:wght@400;600;900&display=swap');
+
+
+$bg-color:rgb(249, 249, 249);
+$bgbox-color:#fff;
+$basic-font-color: #333;
+$point-color: #6eacf5;
+
+
+@mixin flex-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+@mixin flex-space-between {
+  display: flex;
+  justify-content: space-between;
 }
